Bug fix
The bug: When user's mapseries-data repository is not synchronized with the one from moravianlibrary, the application removes and creates a new one with a new name mapseries-data-1 (which is github's standard behavior when deleting and forking the same repository again, when it just existed).
The fix: After the application creates a new repository with a new name, rename it back to mapseries-data.